cernlib 2006
A large collection of general purpose libraries and
modules maintained and offered in both source and
object code form on the CERN central computers. 
Most of these programs were developed at CERN and 
are therefore oriented towards the needs of a physics
research laboratory that is general mathematics,
data analysis, detectors simulation, data-handling
etc... applicable to a wide range of problems.